+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Board: moved SparkFun.lib inside unijoysticle project
+- Firmware: more verbose logs when detecting the type of device
 
 ## [0.2.0] - 2019-05-22
 ### Added